+++ src/components/LogIn.js
@@ -16,6 +16,8 @@ import {
 import { isElectron } from 'utils/electron';
 import { isomorphicReload } from 'utils/isomorphic';
 
+import { ligten } from 'constants/colors';
+
 import Button from 'components/common/controls/Button';
 import Input from 'components/common/controls/Input';
 import Icon from 'components/common/Icon';
@@ -80,6 +82,18 @@ const LoginWrapper = styled.div`
 const StyledGithubLoginButton = styled(Button)`
   width: 200px;
   align-self: center;
+
+  :hover {
+    background: ${(props) => ligten(props.theme.baseAppColor)};
+  }
+`;
+
+const StyledLoginButton = styled(Button)`
+  margin: 0 auto;
+
+  :hover {
+    background: ${(props) => ligten(props.theme.baseAppColor)};
+  }
 `;
 
 const Logo = styled.div`
@@ -264,9 +278,11 @@ export class LogIn extends React.Component {
               <Icon type="info" size="16" color={ this.props.theme.baseAppColor }/>
             </ExternalLink>
             <br/>
-            <Button icon="success" onClick={ () => this.loginWithToken(this.state.fieldsData.token) }>
+            <StyledLoginButton
+              icon="success"
+              onClick={ () => this.loginWithToken(this.state.fieldsData.token) }>
               Log-in
-            </Button>
+            </StyledLoginButton>
             <br/>
             <br/>
             <Anchor onClick={ () => this.setLoginType('basic') }>Cancel</Anchor>
@@ -316,7 +332,7 @@ export class LogIn extends React.Component {
 
             <br/>
             <br/>
-            <Button
+            <StyledLoginButton
               onClick={ () =>
                 this.loginWithBasic(
                   this.state.fieldsData.username,
@@ -326,7 +342,7 @@ export class LogIn extends React.Component {
               }
               icon="success">
               Log-in
-            </Button>
+            </StyledLoginButton>
             <br/>
           </div>
         )}