+++ src/components/layout/headers/SnippetHeader.js
@@ -65,29 +65,6 @@ const Title = styled.div`
   text-overflow: ellipsis;
   position: relative;
   width: 50vw;
-
-  &:after {
-    content: '';
-    width: 100px;
-    height: 50px;
-    position: absolute;
-    top: 3px;
-    right: 0;
-    background: -webkit-gradient(
-      linear,
-      left top,
-      right top,
-      color-stop(0%, rgba(255, 255, 255, 0)),
-      color-stop(56%, rgba(255, 255, 255, 1)),
-      color-stop(100%, rgba(255, 255, 255, 1))
-    );
-    background: -webkit-linear-gradient(
-      left,
-      rgba(255, 255, 255, 0) 0%,
-      rgba(255, 255, 255, 1) 56%,
-      rgba(255, 255, 255, 1) 100%
-    );
-  }
 `;
 
 const Description = styled.span`